+++ src/net/socket.h
@@ -20,7 +20,7 @@ namespace Net
 			void Close();
 			
 			std::string GetIP() const;
-			int GetPort() const;
+			uint16_t GetPort() const;
 			std::string GetAddress() const;
 			std::string GetSocketType() const;
